Implement Part 4.21

+++ part4/bloglist/controller/blogs.js
@@ -63,8 +63,30 @@ blogRouter.put("/:id", async (request, response) => {
   response.status(200).json(updatedBlog);
 });
 
-blogRouter.delete("/:id", async (request, response) => {
+blogRouter.delete("/:id", middleware.extractToken, async (request, response, next) => {
+  let decodedToken;
+
+  try {
+    decodedToken = jwt.verify(request.token, process.env.JWT_SECRET);
+    console.log("Decoded", decodedToken)
+    if (!decodedToken.id) {
+      return response.status(401).json({error: "token invalid"});
+    }
+  } catch (error) {
+    console.log(error)
+    return next(error)
+  }
+
   const id = request.params.id;
+  const blog = await Blog.findById(id);
+  if (!blog) {
+    return response.sendStatus(204);
+  }
+
+  if (blog.user.toString() !== decodedToken.id) {
+    return response.status(403).json({error: "user has no permissions to perform this action"})
+  }
+
   await Blog.findByIdAndDelete(id);
 
   response.sendStatus(204);
